+++ semcore/input/__tests__/input.axe-test.tsx
@@ -3,7 +3,7 @@ import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 
 test.describe('Input', () => {
   test('Basic usage', async ({ page }) => {
-    const standPath = 'website/docs/filter-group/filter-search/examples/dynamic_search.tsx';
+    const standPath = 'stories/patterns/filters/filter-search/docs/examples/dynamic_search.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);