+++ stories/components/breadcrumbs/docs/examples/redefining_a_tag.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Breadcrumbs from '@semcore/breadcrumbs';
+import Link from '@semcore/link';
+
+const Demo = () => (
+  <Breadcrumbs>
+    <Breadcrumbs.Item href='/'>Projects</Breadcrumbs.Item>
+    <Breadcrumbs.Item href='/components/breadcrumbs'>somedomain.com</Breadcrumbs.Item>
+    <Breadcrumbs.Item tag={Link} active href={'#'}>
+      Current page
+    </Breadcrumbs.Item>
+  </Breadcrumbs>
+);
+
+export default Demo;