+++ stories/components/checkbox/docs/examples/additional_props_for_input.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Checkbox from '@semcore/checkbox';
+
+const Demo = () => {
+  return (
+    <Checkbox>
+      <Checkbox.Value>
+        <Checkbox.Value.Control data-testid='checkbox_input_tag' />
+        <Checkbox.Value.CheckMark />
+      </Checkbox.Value>
+      <Checkbox.Text>Value</Checkbox.Text>
+    </Checkbox>
+  );
+};
+
+export default Demo;