+++ stories/components/dropdown-menu/tests/Dropdown-menu.stories.tsx
@@ -1,0 +1,22 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import DropdownMenu from '@semcore/dropdown-menu';
+
+import OnVisibleExample from './examples/on-visible';
+import OnVisible2ndExample from './examples/on-visible-2nd';
+
+const meta: Meta<typeof DropdownMenu> = {
+  title: 'Components/Dropdown Menu/Tests',
+  component: DropdownMenu,
+};
+export default meta;
+
+type Story = StoryObj<typeof DropdownMenu>;
+
+export const OnVisible: Story = {
+  render: OnVisibleExample,
+};
+
+export const OnVisible2nd: Story = {
+    render: OnVisible2ndExample,
+  };