+++ stories/components/input-tags/docs/examples/wrapping_email_in_tag.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import InputTags from '@semcore/input-tags';
-import Select from '@semcore/select';
 import { Text } from '@semcore/typography';
 import { Flex } from '@semcore/flex-box';
 
@@ -42,46 +41,29 @@ const Demo = () => {
     );
   };
 
-  const handleSelect = (value: string) => {
-    changeState([...tags, value], '');
-  };
-
   return (
     <Flex direction='column'>
       <Text tag='label' size={300} htmlFor='add-email'>
         Participants
       </Text>
-      <Select interaction='focus' onChange={handleSelect}>
-        <Select.Trigger
-          tag={InputTags}
-          mt={2}
-          size='l'
-          onAppend={handleAppendTags}
-          onRemove={handleRemoveTag}
-        >
-          {tags.map((tag, idx) => (
-            <InputTags.Tag
-              key={idx}
-              theme='primary'
-              color={isValidEmail(tag) ? 'green-500' : 'red-500'}
-            >
-              <InputTags.Tag.Text>{tag}</InputTags.Tag.Text>
-              <InputTags.Tag.Close data-id={idx} onClick={handleCloseTag} />
-            </InputTags.Tag>
-          ))}
-          <InputTags.Value
-            id='add-email'
-            placeholder='Add email'
-            value={value}
-            onChange={handleChange}
-          />
-        </Select.Trigger>
-        {value && (
-          <Select.Menu>
-            <Select.Option value={value}>{value}</Select.Option>
-          </Select.Menu>
-        )}
-      </Select>
+      <InputTags mt={2} size='l' onAppend={handleAppendTags} onRemove={handleRemoveTag}>
+        {tags.map((tag, idx) => (
+          <InputTags.Tag
+            key={idx}
+            theme='primary'
+            color={isValidEmail(tag) ? 'green-500' : 'red-500'}
+          >
+            <InputTags.Tag.Text>{tag}</InputTags.Tag.Text>
+            <InputTags.Tag.Close data-id={idx} onClick={handleCloseTag} />
+          </InputTags.Tag>
+        ))}
+        <InputTags.Value
+          id='add-email'
+          placeholder='Add email'
+          value={value}
+          onChange={handleChange}
+        />
+      </InputTags>
     </Flex>
   );
 };