+++ stories/components/notice-bubble/docs/examples/special_events_notice.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { NoticeBubbleContainer, NoticeBubbleManager } from '@semcore/notice-bubble';
+import Button from '@semcore/button';
+import MailSent from '@semcore/illustration/MailSent';
+
+const manager = new NoticeBubbleManager();
+
+const Demo = () => {
+  const handleClick = () => {
+    manager.add({
+      children: 'Your post is on its way, and we will take great care of it!',
+      icon: <MailSent />,
+      initialAnimation: true,
+      duration: 10000,
+    });
+  };
+
+  return (
+    <>
+      <Button onClick={handleClick}>Show special event notice</Button>
+      <NoticeBubbleContainer manager={manager} />
+    </>
+  );
+};
+
+export default Demo;