+++ stories/components/select/docs/examples/render_function.tsx
@@ -28,15 +28,25 @@ const Demo = () => (
           getOptionProps, // function encapsulating Select.Option logic
           getOptionCheckboxProps, // function encapsulating Select.Option.Checkbox logic
           value: currentValue, // the current value of the select
+          visible: currentVisible, // the current value of visibility state
         } = props;
         const {
-          visible, // function that controls the internal state of visibility
-          value, // function that controls the internal state of the selected value
+          visible, // function that controls visibility
+          value, // function that controls value
+          highlightedIndex, // function that controls the index of the highlighted item
         } = handlers;
+
+        // we manually highlight the first option from the list except 'select all / deselect all'
+        React.useEffect(() => {
+          if (currentVisible === true) {
+            highlightedIndex(1);
+          }
+        }, [currentVisible]);
 
         const handleClick = () => {
           const newValue = (currentValue as any).length ? [] : options.map(({ value }) => value);
           value(newValue);
+
           return false; // cancel the default handler
         };