+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.47.4] - 2024-11-26
+
+### Fixed
+
+- Keyboard access after changing data in the Table (for example after pagination changes).
+
+## [4.47.3] - 2024-11-22
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.43.2 ~> 4.43.3], `@semcore/core` [2.36.1 ~> 2.36.2]).
+
 ## [4.47.2] - 2024-11-08
 
 ### Changed