+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.43.3] - 2024-11-22
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/dropdown` [4.41.2 ~> 4.41.3], `@semcore/utils` [4.43.2 ~> 4.43.3], `@semcore/core` [2.36.1 ~> 2.36.2]).
+
 ## [4.43.2] - 2024-11-06
 
 ### Fixed