+++ semcore/inline-input/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.44.4] - 2024-11-22
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/flex-box` [5.38.1 ~> 5.38.2], `@semcore/icon` [4.52.1 ~> 4.52.2], `@semcore/utils` [4.43.2 ~> 4.43.3], `@semcore/core` [2.36.1 ~> 2.36.2], `@semcore/button` [5.39.3 ~> 5.39.4]).
+
 ## [4.44.3] - 2024-11-18
 
 ### Changed